+++ tasty-plutus/CHANGELOG.md
@@ -18,8 +18,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ### Added
 
-* `shouldValidateProducing`, `shouldn'tValidateProducing` to extend
-  the functionality of unit testing.
+* Added `outputsToInputs` to perform transformations on the context.
 
 ## 4.1 -- 2021-11-18